Move login step to common-step.js and remove driver.pause method
