+++ wdio/step-definitions/app-launch-times.steps.js
@@ -2,14 +2,10 @@
 import cucumberJson from 'wdio-cucumberjs-json-reporter';
 import { Given, When, Then } from '@wdio/cucumber-framework';
 
-import WalletMainScreen from '../screen-objects/WalletMainScreen';
-import LoginScreen from '../screen-objects/LoginScreen';
-import Accounts from '../helpers/Accounts';
 import WelcomeScreen from '../screen-objects/Onboarding/OnboardingCarousel';
 
 let startTimer;
 let stopTimer;
-const validAccount = Accounts.getValidAccount();
 
 When(/^I kill then app$/, async () => {
   await driver.closeApp();
@@ -18,15 +14,6 @@ When(/^I kill then app$/, async () => {
 When(/^I relaunch the app$/, async () => {
   await driver.startActivity('io.metamask.qa', 'io.metamask.MainActivity');
   startTimer = new Date().getTime();
-});
-
-When(/^I log into my wallet$/, async () => {
-  await WelcomeScreen.waitForSplashAnimationToDisplay();
-  await LoginScreen.isLoginScreenVisible();
-  await LoginScreen.typePassword(validAccount.password);
-  await LoginScreen.tapTitle();
-  await LoginScreen.tapUnlockButton();
-  await WalletMainScreen.isMainWalletViewVisible();
 });
 
 Then(/^the app should launch within x seconds$/, async () => {
