+++ wdio/step-definitions/common-steps.js
@@ -1,5 +1,5 @@
 /* global driver */
-import { Given, Then } from '@wdio/cucumber-framework';
+import { Given, Then, When } from '@wdio/cucumber-framework';
 import Accounts from '../helpers/Accounts';
 import WelcomeScreen from '../screen-objects/Onboarding/OnboardingCarousel';
 import OnboardingScreen from '../screen-objects/Onboarding/OnboardingScreen';
@@ -12,6 +12,7 @@ import CommonScreen from '../screen-objects/CommonScreen';
 
 import SkipAccountSecurityModal from '../screen-objects/Modals/SkipAccountSecurityModal.js';
 import OnboardingWizardModal from '../screen-objects/Modals/OnboardingWizardModal.js';
+import LoginScreen from '../screen-objects/LoginScreen';
 
 Given(/^I have imported my wallet$/, async () => {
   const validAccount = Accounts.getValidAccount();
@@ -64,7 +65,6 @@ Given(/^I import wallet using seed phrase "([^"]*)?"/, async (phrase) => {
   await driver.pause(setTimeout);
   await WelcomeScreen.clickGetStartedButton();
   await OnboardingScreen.clickImportWalletButton();
-  await MetaMetricsScreen.swipeUp();
   await MetaMetricsScreen.tapIAgreeButton();
   const validAccount = Accounts.getValidAccount();
   await ImportFromSeedScreen.typeSecretRecoveryPhrase(phrase);
@@ -112,3 +112,15 @@ Then(
     await WalletMainScreen.isNetworkNameCorrect(networkName);
   },
 );
+
+When(/^I log into my wallet$/, async () => {
+  const validAccount = Accounts.getValidAccount();
+
+  await WelcomeScreen.waitForSplashAnimationToDisplay();
+  await WelcomeScreen.waitForSplashAnimationToDisappear();
+  await LoginScreen.isLoginScreenVisible();
+  await LoginScreen.typePassword(validAccount.password);
+  await LoginScreen.tapTitle();
+  await LoginScreen.tapUnlockButton();
+  await WalletMainScreen.isMainWalletViewVisible();
+});
